Add try_split for safer TX/RX
